refactor: extract file operations module
